WIP xenmgr: Convert vhd support to block-tap script
libxl removed integrated blktap support, but vhds can still be supported
through the block-tap script and tap-ctl.  To do so, we need to change
how the disk specs.

We need to specify "backendtype=phy" to use the block scripts.  We set
"script=block" for true physical devices and "script=block-tap" for vhds
and aio devices.

The "target=" line needs to be prefixed with "vhd:" or "aio:" for the
"black-tap" script, but un-prefixed for real physical devices.

"format" must be "raw" for "backendtype=phy", so that is set
unconditionally.  Failure to do so leads to a libxl parse error.

WIP: This may have broken many things.  But it works for booting NDVM
and UIVM.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>