xenmgr: Set feature-abs-pointer for vkb
We've seen uivm boot up without requesting an absolute pointer -
request-abs-pointer is not present in xenstore.  If feature-abs-pointer
is provided by the backend, linux will write request-abs-pointer.
Hence, it seems uivm starts before the xenstore node is present.

libxl writes some of the xenstore entries, and then vglass writes its
share.  The race could be that the libxl entries are present, but the
vglass ones are not.  Instead of waiting for vglass to write
feature-abs-pointer, have libxl do it.

We'll let libxl write the "height" & "width" as well since those are
subject to the same issue.  "feature-multi-touch" would be there as
well, but we it appears support is broken and wouldn't work for our HVMs
anyway.

This works for vglass, but surfman would need to have a second,
relative pointer added.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>